First cut at being able to omit --phash-distance, --bbox-distance and --track-length
